[wasm] Refactor debug side table
This refactors the debug side table, and extends it to also store
register locations.
In particular
- several vectors are combined into one;
- locals and operand stack values are stored in the same vector;
- three kinds of locations are defined: constant, stack, or register.

When getting the actual values during scope construction, we currently
always report the value -11 for register values. This is to make these
values obvious in the test output. Their value will change once we
actually read the spilled register values from the frame created by the
|WasmDebugBreak| builtin.

R=thibaudm@chromium.org

Bug: v8:10147, v8:10222
Change-Id: I97adb56b31afdb22896530c7ba2e8a24b5d31da9