[parser] Move ParseInfo wrapped argument setup
Change wrapped argument set-up to be closer to where it's needed: setting
up a top-level SFI, or initializing a ParseInfo from a top-level SFI.

This is a generally cleaner use of the interface, avoids splitting the
setting of the funciton syntax kind and wrapped arguments (including
checking script.is_wrapped() in two places for the same behaviour), plus
it avoids unnecessarily creating wrapped_argument handles for functions
inside a wrapped script.

As a drive-by, rename ParseInfo::SetFlagsFromScript to a clearer

ParseInfo::SetFlagsForFunctionInScript, to differentiate between flags
from a script for top-level vs. non-top-level.
Change-Id: Ibdaad957558c13a1528dcc3da1ba8f262f357e48