[wasm][debug] Adding stack scope to CDP and V8
This change adds a stack scope for wasm debugging.
Currently the local scope contains both local variables as well as
the expression stack. For now, this change duplicates the information
available on stacks into the stack scope, until we have added
support for the stack scope in the DevTools front-end.

Bug: chromium: 1043034
Change-Id: Ib0a07e07be7c53003526a7b1e1dbfaa1116b41ad