Update patch set 1
Patch Set 1:

(2 comments)

> Patch Set 1:
> 
> (2 comments)
> 
> One high level comment that this CL brings up, I think it's confusing that all caller's of the CSA Call/TailCall now need to know that JS args are passed in reverse, while CallRuntime/CallStub are still passed on forward order. Would it not be possible to move the reordering down into linkage.cc or somewhere so that it's basically invisible to other code?

Hi Ross,

Thanks for reviewing my CL.

Not all callers of a CSA need to be reversed. I have a new stub in code-assembler.h (CallJSStub) that automatically does this. The only one that needs a bit of care is NewArray.

The runtime function Runtime_NewArray is a bit hacky (check runtime-array.cc). In particular, it has the problem that part of the stack is in the *Runtime Order* and part of it is in the *JS Order*. That's why only part of the arguments in the call is reversed.
Ideally, (I think) we should move Runtime_NewArray to be a builtin (which always have the *JS Order*) and avoid these hacks.

Cheers,

PS: The reason we have two orders (Runtime and JS) is that otherwise we would need to do a massive change in the BytecodeArrayBuilder. It would need to build the bytecode with the arguments already reversed.

Patch-set: 1