[regexp] Split out interface code
This patch moves code at the interface between V8 and irregexp into a
separate file. This code is tightly tied to V8: in particular, it
relies on details of V8 implements garbage collection, strings, and
the interface between native and generated code. By moving this code
into a separate file that SpiderMonkey can omit from its import, it
becomes much easier to keep irregexp in sync between the two
codebases.

R=jgruber@chromium.org

Bug: v8:10303
Change-Id: I9f531a36e4f13440cafb0d0ade921f4c09f39c05