Update patch set 5
Patch Set 5:

Hi Georg,
I found the issue in my CL. The problem was that one kind of call optimization needs to be avoided if the closure belongs to a different native context. And in the cases where we don't have a closure, we can't check that. Therefore, we need to be conservative in that case and avoid that optimization.
Best,
--Michael

Patch-set: 5
Reviewer: Gerrit User 1118564 <1118564@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>