Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

I think this solution should be good enough for now, as it is indeed unclear what kind of solution would be better.

Jakob, can you also take a look? Here is some background:
Bill wants to introduce a builtin with Float64 as parameter 1. The InterfaceDescriptorData does not know anything about FP registers, so it just takes the first default gp register, eax. When we create a CallDescriptor in linkage.cc, we just use the register code in LinkageLocation::ForRegister to create a LinkageLocation. In Bill's case, the register code of eax is the same as for xmm0. The problem now is that xmm0 is not allocatable by the register allocator. This problem caused a crash in the register allocator verifier, and this is the problem Bill tries to solve here.

We thought of the following alternative solutions:
* In linkage.cc in regloc, if it's ia32 and type is a floating point type, then increment the register code by 1. Thereby we avoid to ever allocate xmm0.
  * This will cause problems if someone ever defines an explicit fp register in an interface descriptor. My gut feeling tells me that linkage.cc is not the correct place to fix this issue.
* For any builtin with FP parameters, explicitly define platform-specific registers in {BuiltinName}::InitializePlatformSpecific. 
  * This will make is hard for everyone in the future to add a builtin with FP parameters. Also you have to define registers for all platforms even though you only want to change the register allocation for ia32.
* Change the default registers for ia32 from {eax, ecx, ebx, edx} to {ecx, ebx, edx, eax}, and hope that the forth parameter will never be an FP parameter.
  * This was a good candidate solution. The reason why we decided against it was that eax is also the return register, so builtins that just manipulate the input and return it would need extra register moves. It also bets on the hope that it will not just explode some time later. It would also lead to a crash that is not easy to debug, ask Bill ;-).
* Let InterfaceDescriptorData store not just an array of gp register, but additionally an array of fp registers.
  * This would require to change all the setup machinery of InterfaceDescriptorData to allow to set 2 register arrays instead of one - the whole setup would get more complicated.

Patch-set: 2
Reviewer: Gerrit User 1141154 <1141154@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1