Update patch set 4
Patch Set 4:

Thanks all for the careful consideration and ideas.

I briefly looked at what it would take to make xmm0 allocatable. I believe it's mostly SIMD which invokes instructions which require xmm0 as an implicit operand. But there seem to be many instructions which use xmm0, so this may be a substantial project:

https://source.chromium.org/chromium/chromium/src/+/master:v8/src/codegen/ia32/assembler-ia32.h?originalUrl=https:%2F%2Fcs.chromium.org%2F

Since this is only a problem on ia32, I prefer the work-around that Andreas has suggested, of assigning GP registers in a platform-specific descriptor for the single builtin that I want to add. If our future direction is toward all FP registers being allocatable on ia32, then we'll be able to remove these and I'll leave a TODO there to revisit in that case.

The revised CL just adds a return register and allows returning 1 FP register and 3 GP registers as before. PTAL

Patch-set: 4
Reviewer: Gerrit User 1001213 <1001213@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>