[offthread] Pin OffThreadIsolate to a given thread
Previously OffThreadIsolates set their thread-id on construction. This
thread-id could later be used in DCHECKs, comparing it against the
current thread's id.

However, OffThreadIsolates are created on the main thread (as they need
access to the Isolate and especially Heap for initialization). So, the
thread-id was actually not the background thread's id.

Now, OffThreadIsolate has a PinToCurrentThread method which should be
called on whichever thread wants to actually use it. This pinning can
only be done once, and the OffThreadIsolate is considered invalid before
this method is called.

Bug: chormium:1011762
Change-Id: Ie9d7838152683aea2a326a4e5d1dbd59a747131f