Update patch set 2
Patch Set 2: Code-Review+1

Looks good, I'm only wondering if we should assume that anything with an exponent is automatically a double, similarly to how constants in C are handled.
This is of course assuming that we have that freedom and actual asm.js code will always have 1000 for unsigned and 1.e3 for double and never 1e3. The spec does not seem to exclude the latter explicitly, so this implementation is at least for sure in line with it.

Patch-set: 2
Reviewer: Gerrit User 1381588 <1381588@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1