Update patch set 1
Patch Set 1: Code-Review+1

This rename is OK. A merge conflict there would be easy to resolve.

Related questions which we should discuss offline: Are we going to be able to take advantage of the LocalHeap in the evacuation tasks?

Patch-set: 1
Reviewer: Gerrit User 1119605 <1119605@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1