Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1: Code-Review+1
> > 
> > (2 comments)
> > 
> > LGTM, but I would vote for skipping more generally. Debugging wasm on ppc/s390 would not work anyway, so we could just skip all related tests (even though individual ones might pass).
> 
> I have now skipped all the wasm debug tests.
> 
> Out of curiosity, why are TF debug tests also failing (i.e RunWasmTurbofan_WasmSimpleBreak),
> is debugging fully dependent on liftoff and no fall back to TF?

No, there is no fallback. We cannot set breakpoints in TF, so as soon as you set a breakpoint, that function needs to be compiled in Liftoff. There is a fatal failure if Liftoff bails out in that case.
With the general "tier down on Debugger.enable" thing, there is a fallback, because we only do that for getting scope information. This would be missing if Liftoff is not supported, but we don't fail in that case.

Patch-set: 2
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>