prototype/investigation: profile-guided optimization of builtins
This very sloppy prototype is a proof of concept to see what kind of
performance benefits we might get from applying PGO to builtins. It also
requires a change in chromium.src to make Process::Terminate not
actually terminate the process (so that renderer processes have plenty
of time to finish logging rather than getting terminated by the browser
process).

Design doc:
https://docs.google.com/document/d/1szInbXZfaErWW70d30hJsOLL0Es-l5_g8d2rXm1ZBqI/edit?usp=sharing

Change-Id: I62a2abdfc0b2024a52a4800ca7928ed576b5908e